fix-clippy-0319morning

+++ kernel/src/arch/x86_64/syscall/mod.rs
@@ -152,8 +152,8 @@ pub(super) unsafe fn init_syscall_64() {
     efer |= 0x1;
     x86::msr::wrmsr(x86::msr::IA32_EFER, efer);
 
-    let syscall_base = (1 as u16) << 3;
-    let sysret_base = ((4 as u16) << 3) | 3;
+    let syscall_base = (1_u16) << 3;
+    let sysret_base = ((4_u16) << 3) | 3;
     let high = (u32::from(sysret_base) << 16) | u32::from(syscall_base);
     // 初始化STAR寄存器
     x86::msr::wrmsr(x86::msr::IA32_STAR, u64::from(high) << 32);
